fix: make unit conversions of user provided parameters safe + dont add sample rotation offset twice

+++ packages/essreflectometry/src/ess/estia/conversions.py
@@ -36,7 +36,9 @@ def theta(
     -----------
     The reflection angle of the neutron.
     '''
-    return divergence_angle + sample_rotation.to(unit=divergence_angle.unit)
+    return divergence_angle + sample_rotation.to(
+        unit=divergence_angle.unit, dtype=divergence_angle.dtype
+    )
 
 
 def divergence_angle(
@@ -62,7 +64,9 @@ def divergence_angle(
     The divergence angle of the scattered beam.
     """
     p = position - sample_position.to(unit=position.unit)
-    return sc.atan2(y=p.fields.x, x=p.fields.z) - detector_rotation.to(unit='rad')
+    return sc.atan2(y=p.fields.x, x=p.fields.z) - detector_rotation.to(
+        unit='rad', dtype='float64'
+    )
 
 
 def coordinate_transformation_graph(
